gym_power/requirements.txt
django_rest_framework
requests
django
django_filter
psycopg2-binary
reportlab
openpyxl